Refactoring to clarify which class is a public: move all classes in the ...nio package into ...nio.internal. Move the two new interfaces in ...nio.spi (for TCP extensions) into ...nio. Note that some other gateway projects are explicitly referencing some of the internal classes, mostly in tests in order to inject the required resources into NioSocketAcceptor and NioSocketConnector. The only project referencing an internal class in main code is proxy service, which references NioSocketAcceptor.
