+++ service/proxy/src/test/java/org/kaazing/gateway/service/proxy/ServiceConnectManagerTest.java
@@ -49,7 +49,7 @@ import org.kaazing.gateway.service.ServiceContext;
 import org.kaazing.gateway.transport.BridgeServiceFactory;
 import org.kaazing.gateway.transport.Transport;
 import org.kaazing.gateway.transport.TransportFactory;
-import org.kaazing.gateway.transport.nio.NioSocketAcceptor;
+import org.kaazing.gateway.transport.nio.internal.NioSocketAcceptor;
 import org.kaazing.gateway.util.scheduler.SchedulerProvider;
 import org.slf4j.Logger;
