+++ transport/http/src/test/java/org/kaazing/gateway/transport/http/HttpTransportOptionsTest.java
@@ -52,8 +52,8 @@ import org.kaazing.gateway.transport.BridgeSession;
 import org.kaazing.gateway.transport.TransportFactory;
 import org.kaazing.gateway.transport.test.TransportTestConnectSessionInitializer;
 import org.kaazing.gateway.transport.test.TransportTestIoHandlerAdapter;
-import org.kaazing.gateway.transport.nio.NioSocketAcceptor;
-import org.kaazing.gateway.transport.nio.NioSocketConnector;
+import org.kaazing.gateway.transport.nio.internal.NioSocketAcceptor;
+import org.kaazing.gateway.transport.nio.internal.NioSocketConnector;
 import org.kaazing.gateway.util.scheduler.SchedulerProvider;
 import org.kaazing.mina.core.buffer.IoBufferAllocatorEx;
 import org.kaazing.mina.core.future.UnbindFuture;
