+++ transport/nio/src/main/java/org/kaazing/gateway/transport/nio/TcpExtensionFactorySpi.java
@@ -19,7 +19,7 @@
  * under the License.
  */
 
-package org.kaazing.gateway.transport.nio.spi;
+package org.kaazing.gateway.transport.nio;
 
 import org.kaazing.gateway.resource.address.ResourceAddress;
