+++ transport/nio/src/main/java/org/kaazing/gateway/transport/nio/internal/NioSocketConnector.java
@@ -19,18 +19,18 @@
  * under the License.
  */
 
-package org.kaazing.gateway.transport.nio;
+package org.kaazing.gateway.transport.nio.internal;
 
 import static java.util.concurrent.Executors.newCachedThreadPool;
-import static org.kaazing.gateway.transport.nio.NioSystemProperty.TCP_IP_TOS;
-import static org.kaazing.gateway.transport.nio.NioSystemProperty.TCP_KEEP_ALIVE;
-import static org.kaazing.gateway.transport.nio.NioSystemProperty.TCP_MAXIMUM_READ_BUFFER_SIZE;
-import static org.kaazing.gateway.transport.nio.NioSystemProperty.TCP_MINIMUM_READ_BUFFER_SIZE;
-import static org.kaazing.gateway.transport.nio.NioSystemProperty.TCP_NO_DELAY;
-import static org.kaazing.gateway.transport.nio.NioSystemProperty.TCP_READ_BUFFER_SIZE;
-import static org.kaazing.gateway.transport.nio.NioSystemProperty.TCP_RECEIVE_BUFFER_SIZE;
-import static org.kaazing.gateway.transport.nio.NioSystemProperty.TCP_SEND_BUFFER_SIZE;
-import static org.kaazing.gateway.transport.nio.NioSystemProperty.TCP_SO_LINGER;
+import static org.kaazing.gateway.transport.nio.internal.NioSystemProperty.TCP_IP_TOS;
+import static org.kaazing.gateway.transport.nio.internal.NioSystemProperty.TCP_KEEP_ALIVE;
+import static org.kaazing.gateway.transport.nio.internal.NioSystemProperty.TCP_MAXIMUM_READ_BUFFER_SIZE;
+import static org.kaazing.gateway.transport.nio.internal.NioSystemProperty.TCP_MINIMUM_READ_BUFFER_SIZE;
+import static org.kaazing.gateway.transport.nio.internal.NioSystemProperty.TCP_NO_DELAY;
+import static org.kaazing.gateway.transport.nio.internal.NioSystemProperty.TCP_READ_BUFFER_SIZE;
+import static org.kaazing.gateway.transport.nio.internal.NioSystemProperty.TCP_RECEIVE_BUFFER_SIZE;
+import static org.kaazing.gateway.transport.nio.internal.NioSystemProperty.TCP_SEND_BUFFER_SIZE;
+import static org.kaazing.gateway.transport.nio.internal.NioSystemProperty.TCP_SO_LINGER;
 
 import java.util.Properties;
