+++ transport/nio/src/main/java/org/kaazing/gateway/transport/nio/internal/UdpTransport.java
@@ -19,7 +19,7 @@
  * under the License.
  */
 
-package org.kaazing.gateway.transport.nio;
+package org.kaazing.gateway.transport.nio.internal;
 
 import java.util.Collections;
 import java.util.HashMap;
