+++ transport/nio/src/main/resources/META-INF/services/org.kaazing.gateway.transport.TransportFactorySpi
@@ -1,2 +1,2 @@
-org.kaazing.gateway.transport.nio.TcpTransportFactorySpi
-org.kaazing.gateway.transport.nio.UdpTransportFactorySpi
+org.kaazing.gateway.transport.nio.internal.TcpTransportFactorySpi
+org.kaazing.gateway.transport.nio.internal.UdpTransportFactorySpi
