+++ transport/nio/src/test/java/org/kaazing/gateway/transport/nio/internal/NioTransportTest.java
@@ -19,35 +19,14 @@
  * under the License.
  */
 
-package org.kaazing.gateway.transport.nio;
+package org.kaazing.gateway.transport.nio.internal;
 
-import java.util.Collection;
-import java.util.Collections;
-import java.util.Map;
-import java.util.Properties;
+import org.junit.Test;
 
-import org.kaazing.gateway.transport.Transport;
-import org.kaazing.gateway.transport.TransportFactorySpi;
+public class NioTransportTest {
 
-public final class UdpTransportFactorySpi extends TransportFactorySpi {
-
-    private final Collection<String> UDP_SCHEMES = Collections.singleton("udp");
-
-    @Override
-    public String getTransportName() {
-        return "udp";
+    @Test 
+    public void test() throws Exception {
     }
-
-    @Override
-    public Collection<String> getSchemeNames() {
-        return UDP_SCHEMES;
-    }
-
-    @Override
-    public Transport newTransport(Map<String, ?> configuration) {
-        Properties properties = new Properties();
-        properties.putAll(configuration);
-        return new UdpTransport(properties);
-    }
-
+    
 }
