+++ transport/wsn/src/test/java/org/kaazing/gateway/transport/wsn/WsnBindingsTest.java
@@ -43,8 +43,8 @@ import org.kaazing.gateway.transport.IoHandlerAdapter;
 import org.kaazing.gateway.transport.TransportFactory;
 import org.kaazing.gateway.transport.http.HttpAcceptor;
 import org.kaazing.gateway.transport.http.HttpConnector;
-import org.kaazing.gateway.transport.nio.NioSocketAcceptor;
-import org.kaazing.gateway.transport.nio.NioSocketConnector;
+import org.kaazing.gateway.transport.nio.internal.NioSocketAcceptor;
+import org.kaazing.gateway.transport.nio.internal.NioSocketConnector;
 import org.kaazing.gateway.transport.ssl.SslAcceptor;
 import org.kaazing.gateway.transport.ssl.bridge.filter.SslCertificateSelectionFilter;
 import org.kaazing.gateway.transport.ws.WsAcceptor;
