+++ transport/wsn/src/test/java/org/kaazing/gateway/transport/wsn/WsnTransportTest.java
@@ -46,8 +46,8 @@ import org.kaazing.gateway.transport.IoHandlerAdapter;
 import org.kaazing.gateway.transport.TransportFactory;
 import org.kaazing.gateway.transport.http.HttpAcceptor;
 import org.kaazing.gateway.transport.http.HttpConnector;
-import org.kaazing.gateway.transport.nio.NioSocketAcceptor;
-import org.kaazing.gateway.transport.nio.NioSocketConnector;
+import org.kaazing.gateway.transport.nio.internal.NioSocketAcceptor;
+import org.kaazing.gateway.transport.nio.internal.NioSocketConnector;
 import org.kaazing.gateway.transport.ws.WsAcceptor;
 import org.kaazing.gateway.transport.ws.bridge.filter.WsBufferAllocator;
 import org.kaazing.gateway.transport.ws.extension.WebSocketExtensionFactory;
