+++ transport/wsr/src/test/java/org/kaazing/gateway/transport/wsr/WsrTransportTest.java
@@ -41,8 +41,8 @@ import org.kaazing.gateway.transport.BridgeSession;
 import org.kaazing.gateway.transport.TransportFactory;
 import org.kaazing.gateway.transport.http.HttpAcceptor;
 import org.kaazing.gateway.transport.http.HttpConnector;
-import org.kaazing.gateway.transport.nio.NioSocketAcceptor;
-import org.kaazing.gateway.transport.nio.NioSocketConnector;
+import org.kaazing.gateway.transport.nio.internal.NioSocketAcceptor;
+import org.kaazing.gateway.transport.nio.internal.NioSocketConnector;
 import org.kaazing.gateway.transport.pipe.NamedPipeAcceptor;
 import org.kaazing.gateway.transport.pipe.NamedPipeConnector;
 import org.kaazing.gateway.transport.ws.WsAcceptor;
